requirements.txt
boto3
pygithub
